+++ src/server/CHANGELOG.md
@@ -1,5 +1,10 @@
 # SAMS Changelog
 
+## 0.0.1.dev2
+- [SDESK-5287] Set resource and service
+
+## 0.0.1.dev1
+- [SDESK-5285] Create skeleton application structure
+
 ## June 2020
 - Initial commit
-- [SDESK-5285] Create skeleton application structure